Cập nhật cấu trúc UI: tách thư mục group_chat, post, profile và xóa view cũ
