--- lib/view/chat_view.dart
@@ -1,186 +1,0 @@
-// lib/view/chat_view.dart
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:mangxahoi/viewmodel/chat_viewmodel.dart';
-import 'package:mangxahoi/model/model_message.dart';
-import 'package:mangxahoi/model/model_user.dart';
-import 'package:mangxahoi/authanet/firestore_listener.dart';
-import 'package:intl/intl.dart';
-
-class ChatView extends StatelessWidget {
-  final String chatId;
-  final String chatName;
-
-  const ChatView({super.key, required this.chatId, required this.chatName});
-
-  @override
-  Widget build(BuildContext context) {
-    return ChangeNotifierProvider(
-      create: (_) => ChatViewModel(chatId: chatId),
-      child: _ChatViewContent(chatName: chatName),
-    );
-  }
-}
-
-class _ChatViewContent extends StatelessWidget {
-  final String chatName;
-  const _ChatViewContent({required this.chatName});
-
-  @override
-  Widget build(BuildContext context) {
-    final vm = context.watch<ChatViewModel>();
-    final firestoreListener = context.watch<FirestoreListener>();
-
-    return Scaffold(
-      appBar: AppBar(title: Text(chatName)),
-      body: Column(
-        children: [
-          Expanded(
-            child: StreamBuilder<List<MessageModel>>(
-              stream: vm.messagesStream,
-              builder: (context, snapshot) {
-                if (!snapshot.hasData) {
-                  return const Center(child: CircularProgressIndicator());
-                }
-                final messages = snapshot.data!;
-                return ListView.builder(
-                  reverse: true,
-                  padding: const EdgeInsets.all(10.0),
-                  itemCount: messages.length,
-                  itemBuilder: (context, index) {
-                    final message = messages[index];
-                    final sender = firestoreListener.getUserByAuthUid(message.senderId);
-                    final isMe = message.senderId == vm.currentUserId;
-                    return _MessageBubble(
-                      message: message,
-                      sender: sender,
-                      isMe: isMe,
-                    );
-                  },
-                );
-              },
-            ),
-          ),
-          _buildMessageComposer(context, vm),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildMessageComposer(BuildContext context, ChatViewModel vm) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 8.0),
-      height: 70.0,
-      color: Colors.white,
-      child: Row(
-        children: <Widget>[
-          Expanded(
-            child: TextField(
-              controller: vm.messageController,
-              textCapitalization: TextCapitalization.sentences,
-              decoration: const InputDecoration.collapsed(
-                hintText: 'Nhập tin nhắn...',
-              ),
-            ),
-          ),
-          IconButton(
-            icon: const Icon(Icons.send),
-            iconSize: 25.0,
-            color: Theme.of(context).primaryColor,
-            onPressed: vm.sendMessage,
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// Widget mới để hiển thị tin nhắn chi tiết
-class _MessageBubble extends StatelessWidget {
-  final MessageModel message;
-  final UserModel? sender;
-  final bool isMe;
-
-  const _MessageBubble({
-    required this.message,
-    required this.sender,
-    required this.isMe,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final Radius messageRadius = const Radius.circular(20);
-    final avatarImage = sender?.avatar.isNotEmpty ?? false ? NetworkImage(sender!.avatar.first) : null;
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 6.0),
-      child: Row(
-        mainAxisAlignment: isMe ? MainAxisAlignment.end : MainAxisAlignment.start,
-        crossAxisAlignment: CrossAxisAlignment.end,
-        children: <Widget>[
-          if (!isMe) ...[
-            CircleAvatar(
-              radius: 15.0,
-              backgroundImage: avatarImage,
-              child: avatarImage == null ? const Icon(Icons.person, size: 15) : null,
-            ),
-            const SizedBox(width: 8.0),
-          ],
-          Flexible(
-            child: Column(
-              crossAxisAlignment: isMe ? CrossAxisAlignment.end : CrossAxisAlignment.start,
-              children: [
-                if (!isMe && sender != null)
-                  Padding(
-                    padding: const EdgeInsets.only(left: 10.0, bottom: 2.0),
-                    child: Text(
-                      sender!.name,
-                      style: const TextStyle(fontSize: 12.0, color: Colors.black54),
-                    ),
-                  ),
-                Container(
-                  constraints: BoxConstraints(
-                    maxWidth: MediaQuery.of(context).size.width * 0.75,
-                  ),
-                  padding: const EdgeInsets.symmetric(horizontal: 15.0, vertical: 10.0),
-                  decoration: BoxDecoration(
-                    color: isMe ? Theme.of(context).primaryColor : Colors.grey[200],
-                    borderRadius: isMe
-                        ? BorderRadius.only(
-                            topLeft: messageRadius,
-                            bottomLeft: messageRadius,
-                            topRight: messageRadius,
-                          )
-                        : BorderRadius.only(
-                            topRight: messageRadius,
-                            bottomRight: messageRadius,
-                            topLeft: messageRadius,
-                          ),
-                  ),
-                  child: Text(
-                    message.content,
-                    style: TextStyle(
-                      color: isMe ? Colors.white : Colors.black,
-                      fontSize: 15.0,
-                    ),
-                  ),
-                ),
-                const SizedBox(height: 4),
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 10.0),
-                  child: Text(
-                    DateFormat('HH:mm').format(message.createdAt),
-                    style: const TextStyle(
-                      fontSize: 10.0,
-                      color: Colors.grey,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
